Исправление пустого имени для человека зашедшего через номер телефона

+++ app/src/main/java/live/noxbox/Configuration.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public interface Configuration {
     int REQUESTING_AND_ACCEPTING_TIMEOUT_IN_SECONDS = 60;
-    int REQUESTING_AND_ACCEPTING_TIMEOUT_IN_MILLIS = 60 * 1000;
+    long REQUESTING_AND_ACCEPTING_TIMEOUT_IN_MILLIS = 60 * 1000;
     String START_TIME = "00:00:00";
     double RADIUS_IN_METERS = 50 * 1000;
     int MIN_RATE_IN_PERCENTAGE = 90;
